Migrate test & demo projects to new setup format
