+++ demo-projects/blog/index.js
@@ -1,6 +1,3 @@
-//imports for Next.js app core
-const next = require('next');
-
 //imports for Keystone app core
 const { AdminUI } = require('@voussoir/admin-ui');
 const { Keystone } = require('@voussoir/keystone');
@@ -14,41 +11,14 @@ const {
   CalendarDay,
   DateTime,
 } = require('@voussoir/fields');
-const { WebServer } = require('@voussoir/server');
 const { LocalFileAdapter } = require('@voussoir/file-adapters');
 const PasswordAuthStrategy = require('@voussoir/keystone/auth/Password');
 const { MongooseAdapter } = require('@voussoir/adapter-mongoose');
 
 // config
 const path = require('path');
-const port = 3000;
 const staticRoute = '/public'; // The URL portion
 const staticPath = path.join(process.cwd(), 'public'); // The local path on disk
-const initialData = {
-  User: [
-    {
-      name: 'Administrator',
-      email: 'admin@keystone.project',
-      isAdmin: true,
-      dob: '1990-01-01',
-      password: 'password',
-    },
-    {
-      name: 'Demo User',
-      email: 'a@demo.user',
-      isAdmin: false,
-      dob: '1995-06-09',
-      password: 'password',
-    },
-  ],
-};
-
-const nextApp = next({
-  dir: 'app',
-  distDir: 'build',
-  dev: process.env.NODE_ENV !== 'PRODUCTION',
-});
-
 const getYear = require('date-fns/get_year');
 
 const keystone = new Keystone({
@@ -146,30 +116,10 @@ const admin = new AdminUI(keystone, {
   sortListsAlphabetically: true,
 });
 
-const server = new WebServer(keystone, {
-  'cookie secret': 'qwerty',
-  'admin ui': admin,
-  port,
-  authStrategy: authStrategy,
-});
-
-server.app.use(staticRoute, server.express.static(staticPath));
-
-server.app.use(nextApp.getRequestHandler());
-
-async function start() {
-  await Promise.all([keystone.connect(), nextApp.prepare()]);
-  server.start();
-  const users = await keystone.lists.User.adapter.findAll();
-  if (!users.length) {
-    Object.values(keystone.adapters).forEach(async adapter => {
-      await adapter.dropDatabase();
-    });
-    await keystone.createItems(initialData);
-  }
-}
-
-start().catch(error => {
-  console.error(error);
-  process.exit(1);
-});
+module.exports = {
+  staticRoute,
+  staticPath,
+  keystone,
+  authStrategy,
+  admin,
+};
