+++ demo-projects/blog/package.json
@@ -8,7 +8,7 @@
     "node": ">=8.4.0"
   },
   "scripts": {
-    "start": "DISABLE_LOGGING=true node index.js",
+    "start": "DISABLE_LOGGING=true keystone",
     "build": "next build"
   },
   "dependencies": {
@@ -16,9 +16,10 @@
     "@emotion/styled": "^10.0.4",
     "@voussoir/adapter-mongoose": "^2.0.0",
     "@voussoir/admin-ui": "^1.0.0",
-    "@voussoir/keystone": "^0.0.0",
+    "@voussoir/core": "^2.1.0",
     "@voussoir/fields": "^3.0.0",
     "@voussoir/file-adapters": "^0.2.1",
+    "@voussoir/keystone": "^0.0.0",
     "@voussoir/server": "^1.0.0",
     "apollo-cache-inmemory": "^1.5.1",
     "apollo-client": "^2.5.1",
