+++ demo-projects/todo/index.js
@@ -1,11 +1,7 @@
-// Keystone imports
 const { Keystone } = require('@voussoir/keystone');
 const { AdminUI } = require('@voussoir/admin-ui');
-const { Text } = require('@voussoir/fields');
-const { WebServer } = require('@voussoir/server');
 const { MongooseAdapter } = require('@voussoir/adapter-mongoose');
-const express = require('express');
-const path = require('path');
+const { Text } = require('@voussoir/fields');
 
 const keystone = new Keystone({
   name: 'Keystone To-Do List',
@@ -19,22 +15,10 @@ keystone.createList('Todo', {
   },
 });
 
-const admin = new AdminUI(keystone, { adminPath: '/admin' });
+// Setup the optional Admin UI
+const admin = new AdminUI(keystone);
 
-const server = new WebServer(keystone, {
-  'cookie secret': 'qwerty',
-  'admin ui': admin,
-  port: 3000,
-});
-
-server.app.use(express.static(path.join(__dirname, 'public')));
-
-async function start() {
-  await keystone.connect();
-  server.start();
-}
-
-start().catch(error => {
-  console.error(error);
-  process.exit(1);
-});
+module.exports = {
+  keystone,
+  admin,
+};
