+++ test-projects/basic/cypress/mock/upload.txt
@@ -1,1 +1,1 @@
-Some important content 0.6477825929662109
+Some important content 0.7420151887341895
