+++ test-projects/basic/index.js
@@ -16,10 +16,9 @@ const {
   Content,
 } = require('@voussoir/fields');
 const Decimal = require('../../packages/fields/types/Decimal');
-const { WebServer } = require('@voussoir/server');
 const { CloudinaryAdapter, LocalFileAdapter } = require('@voussoir/file-adapters');
 
-const { port, staticRoute, staticPath, cloudinary } = require('./config');
+const { staticRoute, staticPath, cloudinary } = require('./config');
 
 const LOCAL_FILE_PATH = `${staticPath}/avatars`;
 const LOCAL_FILE_ROUTE = `${staticRoute}/avatars`;
@@ -29,8 +28,6 @@ const getYear = require('date-fns/get_year');
 
 // TODO: Make this work again
 // const SecurePassword = require('./custom-fields/SecurePassword');
-
-const initialData = require('./data');
 
 const { MongooseAdapter } = require('@voussoir/adapter-mongoose');
 
@@ -161,38 +158,7 @@ const admin = new AdminUI(keystone, {
   sortListsAlphabetically: true,
 });
 
-const server = new WebServer(keystone, {
-  'cookie secret': 'qwerty',
-  'admin ui': admin,
-  port,
-});
-
-server.app.get('/reset-db', (req, res) => {
-  const reset = async () => {
-    Object.values(keystone.adapters).forEach(async adapter => {
-      await adapter.dropDatabase();
-    });
-    await keystone.createItems(initialData);
-    res.redirect(admin.adminPath);
-  };
-  reset();
-});
-
-server.app.use(staticRoute, server.express.static(staticPath));
-
-async function start() {
-  await keystone.connect(process.env.MONGODB_URI);
-  server.start();
-  const users = await keystone.lists.User.adapter.findAll();
-  if (!users.length) {
-    Object.values(keystone.adapters).forEach(async adapter => {
-      await adapter.dropDatabase();
-    });
-    await keystone.createItems(initialData);
-  }
-}
-
-start().catch(error => {
-  console.error(error);
-  process.exit(1);
-});
+module.exports = {
+  keystone,
+  admin,
+};
